refactor: move Scanner to its own file, remove Files.scan
