+++ ggshield/cmd/secret/scan/pypi.py
@@ -15,7 +15,7 @@ from ggshield.core.file_utils import get_files_from_paths
 from ggshield.core.text_utils import create_progress_bar
 from ggshield.core.utils import ScanContext, ScanMode
 from ggshield.output import OutputHandler
-from ggshield.scan import Files, ScanCollection
+from ggshield.scan import Files, ScanCollection, Scanner
 
 
 PYPI_DOWNLOAD_TIMEOUT = 30
@@ -110,12 +110,15 @@ def pypi_cmd(ctx: click.Context, package_name: str) -> int:  # pragma: no cover
                 command_path=ctx.command_path,
             )
 
-            results = files.scan(
+            scanner = Scanner(
                 client=ctx.obj["client"],
                 cache=ctx.obj["cache"],
-                matches_ignore=config.secret.ignored_matches,
+                ignored_matches=config.secret.ignored_matches,
                 scan_context=scan_context,
                 ignored_detectors=config.secret.ignored_detectors,
+            )
+            results = scanner.scan(
+                files.files,
                 progress_callback=partial(progress.update, task_scan),
                 scan_threads=MAX_WORKERS,
             )
