+++ tests/unit/scan/test_scan.py
@@ -6,7 +6,7 @@ from click import Command, Context, Group
 from ggshield import __version__
 from ggshield.core.cache import Cache
 from ggshield.core.utils import ScanContext, ScanMode
-from ggshield.scan import Commit
+from ggshield.scan import Commit, Scanner
 from ggshield.scan.repo import cd
 from tests.unit.conftest import UNCHECKED_SECRET_PATCH
 
@@ -26,15 +26,15 @@ def test_request_headers(scan_mock: Mock, client):
 
     with Context(Command("bar"), info_name="bar") as ctx:
         ctx.parent = Context(Group("foo"), info_name="foo")
-        c.scan(
+        scanner = Scanner(
             client=client,
             cache=Cache(),
-            matches_ignore={},
             scan_context=ScanContext(
                 scan_mode=ScanMode.PATH,
                 command_path=ctx.command_path,
             ),
         )
+        scanner.scan(c.files)
     scan_mock.assert_called_with(
         ANY,
         {
